ifcfg-rh: fix reading SR-IOV settings
Fixes: d48f389cbf6a24f6e58f1422e5b5a8a8d649c973

+++ src/settings/plugins/ifcfg-rh/nms-ifcfg-rh-reader.c
@@ -2286,7 +2286,7 @@ make_sriov_setting (shvarFile *ifcfg)
 	total_vfs = NM_MAX (total_vfs, 0);
 
 	g_object_set (s_sriov,
-	              NM_SETTING_SRIOV_TOTAL_VFS, total_vfs,
+	              NM_SETTING_SRIOV_TOTAL_VFS, (guint) total_vfs,
 	              NM_SETTING_SRIOV_VFS, vfs,
 	              NM_SETTING_SRIOV_AUTOPROBE_DRIVERS, autoprobe_drivers,
 	              NULL);
